refactor: CodeBuildForTagBuildを分離する。
